Bug fix
Fixes a minor bug that can occur when multiple TEs/cCREs overlap a coordinate. This can happen in a few instances, such as when overlapping TE annotations are present and when there are multiple liftOver coordinates (potentially caused by deletions/expansions). Only takes the first TE/cCRE that for output.